New docs do not require fake components, all typing files should be types.ts
