+++ src/components/loaderScreen/types.ts
@@ -1,4 +1,3 @@
-import {PureComponent} from 'react';
 import {ActivityIndicatorProps, TextStyle, ViewStyle} from 'react-native';
 
 export type LoaderScreenProps = ActivityIndicatorProps & {
@@ -31,17 +30,3 @@ export type LoaderScreenProps = ActivityIndicatorProps & {
      */
     containerStyle?: ViewStyle;
 };
-
-
-/**
- * @description: Component that shows a full screen with an activity indicator
- * @example: https://github.com/wix/react-native-ui-lib/blob/master/demo/src/screens/componentScreens/LoadingScreen.tsx
- */
-// @ts-ignore
-class FakeLoaderScreenForDocs extends PureComponent<LoaderScreenProps> { // eslint-disable-line
-  static displayName = 'LoaderScreen';
-
-  render() {
-    return null;
-  }
-}
