+++ src/components/stateScreen/types.ts
@@ -1,4 +1,3 @@
-import {PureComponent} from 'react';
 import {ImageURISource} from 'react-native';
 
 export type StateScreenProps = {
@@ -29,17 +28,3 @@ export type StateScreenProps = {
   testId?: string;
   testID?: string;
 };
-
-/**
- * @description: Component that shows a full screen for a certain state, like an empty state
- * @image: https://user-images.githubusercontent.com/33805983/34672894-f262ab84-f488-11e7-83f0-4ee0f0ac34ba.png
- * @example: https://github.com/wix/react-native-ui-lib/blob/master/demo/src/screens/componentScreens/EmptyStateScreen.tsx
- */
-// @ts-ignore
-class FakeStateScreenForDocs extends PureComponent<StateScreenProps> { // eslint-disable-line
-  static displayName = 'StateScreen';
-
-  render() {
-    return null;
-  }
-}
